Towards a fix for VOMS-488
VOMS Admin now allows users to change their email also when the OrgDB
integration is enabled. The code checks that the user name and surname
in the VOMS database are equal to the name and surname in the OrgDB.

Now VOMS Admin automatically restores memberships that were suspended
due to missing records in the OrgDB (this happens when there is no
match between the email address in VOMS and in the OrgDB).