Fix VOMS-608
An error in the structure of the generated caused errors when multiple
pending VO membership requests where present. It was possible to approve
one request, but then the approve/reject buttons would no longer work.
This was caused by the incorrect nesting between the html form and
table elements. HTML mandates that a form contains the whole table or
is contained in a table cell. form cannot contain fragments of a table.

Issue: https://issues.infn.it/jira/browse/VOMS-608