Initial work towards a complete audit log
In order to have a complete audit log, all actions that change the VOMS
database state and could affect VO membership and structure need to be
recorded in the database.

We leverage the existing event system (sligthly refactored) and plug a
listener that logs an audit entry in the database whenever "interesting"
events are raised.

The audit entry is part of the same transaction of the other persistence
operations.

This change affects much of the VOMS Admin logic as we must insert
the trigger of interesting events wherever it makes sense.