Handle unconfirmed request as other requests
Previously VOMS Admin did not allow the approval/rejection of
unconfirmed requests. If a VO admin wanted to approve an unconfirmed
request he would have to first confirm it on behalf of the user and
then approve it. Now this two steps are combined in a single operation
and unconfirmed requests are treated as other requests.