Don't expect ojdbc6.jar in VOMS Admin server libs
As we are not going to ship oracle jdbc jar in VOMS Admin RPMs anymore
(this is long due), the startup command has been changed to take it from
where the sysadmin will want it to be.

The VOMS Admin init script uses the OJDBC_JAR variable to locate the
oracle jar to be included in the classpath.

If OJDBC_JAR is not explicitly defined in

/etc/sysconfig/voms-admin

it will be set to the following value:

$ORACLE_LIBRARY_PATH/ojdbc6.jar

with ORACLE_LIBRARY_PATH as already defined in
/etc/sysconfig/voms-admin.