Tentative fix for VOMS-704
An experiment participation in the CERN HR db can be open-ended, i.e.
have a start date but not an end date. Currently VOMS extends the
membership lifetime for a year 5 days before membership expiration.
It's safer to extend the membership end time each time the synch task
is run.