Synchronization for background tasks
This commit introduces a locking mechanism, implemented using a
dedicated database table, to avoid races in the concurrent execution of
background tasks that change the VOMS database contents.

This mechanism is useful for deployments that run multiple master
instances of the voms-admin services, in order to avoid unnecessary
execution of background task in separate instances or conflicts that may
arise if background task run concurrently (like for instance, multiple
sign aup tasks assigned to users).

Each task at start tries to acquire a lock from the task_lock table.
The lock is acquired if not other service is running the same task (or
has already ran the task in the time period configured for the task).