Work around commons-io jar hell
CANL wants commons-io-1.3.2.

The latest struts, and in particular the fileupload functionality,
requires commons-io-2.2. With 1.3.2 fileupload fails (some class looks
for a method that is not available in 1.3.2).

When using 2.2, CANL connector startup fails, since a method CANL relies
on was removed in 2.2.

So the workaround is to use commons-io-1.3.2 for the container and
commons-io-2.2 inside the voms-admin war, which however required some
gymnastics in the build configuration.